# SWC: Rust based compiler to replace babel