--- a/packages/magento-cart-billing-address/components/EditBillingAddressForm/EditBillingAddressForm.tsx
+++ b/packages/magento-cart-billing-address/components/EditBillingAddressForm/EditBillingAddressForm.tsx
@@ -16,7 +16,8 @@ import {
   InputCheckmark,
 } from '@graphcommerce/next-ui'
 import { phonePattern } from '@graphcommerce/react-hook-form'
-import { t, Trans } from '@lingui/macro'
+import { i18n } from '@lingui/core'
+import { Trans } from '@lingui/react'
 import { SxProps, TextField, Theme } from '@mui/material'
 import { GetBillingAddressDocument } from './GetBillingAddress.gql'
 import { SetBillingAddressOnCartDocument } from './SetBillingAddressOnCart.gql'
@@ -71,10 +72,10 @@ export function EditBillingAddressForm(props: EditBillingAddressFormProps) {
             type='text'
             error={!!formState.errors.telephone}
             required={required.telephone}
-            label={<Trans>Telephone</Trans>}
+            label={<Trans id='Telephone' />}
             {...muiRegister('telephone', {
               required: required.telephone,
-              pattern: { value: phonePattern, message: t`Invalid phone number` },
+              pattern: { value: phonePattern, message: i18n._(/* i18n */ `Invalid phone number`) },
             })}
             helperText={formState.isSubmitted && formState.errors.telephone?.message}
             disabled={formState.isSubmitting}
@@ -92,7 +93,7 @@ export function EditBillingAddressForm(props: EditBillingAddressFormProps) {
             size='large'
             loading={formState.isSubmitting}
           >
-            <Trans>Save changes</Trans>
+            <Trans id='Save changes' />
           </Button>
         </FormActions>
       </Form>