--- a/packages/magento-cart/hooks/useCartIdCreate.ts
+++ b/packages/magento-cart/hooks/useCartIdCreate.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@graphcommerce/graphql'
-import { t } from '@lingui/macro'
+import { i18n } from '@lingui/core'
 import { CreateEmptyCartDocument } from './CreateEmptyCart.gql'
 import { useAssignCurrentCartId } from './useAssignCurrentCartId'
 import { useCurrentCartId } from './useCurrentCartId'
@@ -13,7 +13,7 @@ export function useCartIdCreate() {
     if (cartId) return cartId
 
     const { data } = await create()
-    if (!data?.createEmptyCart) throw Error(t`Could not create an empty cart`)
+    if (!data?.createEmptyCart) throw Error(i18n._(/* i18n */ `Could not create an empty cart`))
 
     // We store the cartId that is returned as the currentCartId result
     assignCurrentCartId(data.createEmptyCart)