… .changeset/little-trees-clap.md
… .changeset/ten-cameras-talk.md
… .changeset/tricky-spiders-reflect.md
… docs/framework/translations.md
… examples/magento-graphcms/.babelrc
… examples/magento-graphcms/components/Blog/BlogItem.tsx
… examples/magento-graphcms/components/Layout/Footer.tsx
… examples/magento-graphcms/components/Layout/LayoutFull.tsx
… examples/magento-graphcms/locales/de.po
… examples/magento-graphcms/locales/en.po
… examples/magento-graphcms/locales/es.po
… examples/magento-graphcms/locales/fr.po
… examples/magento-graphcms/locales/it.po
… examples/magento-graphcms/locales/nl.po
… examples/magento-graphcms/next.config.js
… examples/magento-graphcms/package.json
… examples/magento-graphcms/pages/404.tsx
… examples/magento-graphcms/pages/account/addresses/add.tsx
… examples/magento-graphcms/pages/account/addresses/edit.tsx
… examples/magento-graphcms/pages/account/addresses/index.tsx
… examples/magento-graphcms/pages/account/authentication/index.tsx
… examples/magento-graphcms/pages/account/contact/index.tsx
… examples/magento-graphcms/pages/account/forgot-password.tsx
… examples/magento-graphcms/pages/account/index.tsx
… examples/magento-graphcms/pages/account/name/index.tsx
… examples/magento-graphcms/pages/account/orders/index.tsx
… examples/magento-graphcms/pages/account/orders/view.tsx
… examples/magento-graphcms/pages/account/reviews/add.tsx
… examples/magento-graphcms/pages/account/reviews/index.tsx
… examples/magento-graphcms/pages/account/signin.tsx