With this PR, GraphCommerce will be using the Rust based [SWC](https://swc.rs/) compiler that will replace [babel](https://babeljs.io/). With this GraphCommerce is embracing that [Rust is the future of javascript](https://leerob.io/blog/rust).

SWC was released as part of [Next.js v12](https://nextjs.org/blog/next-12#faster-builds-and-fast-refresh-with-rust-compiler) and has been stabilizing in the last few months. With the [v12.1.1 release](https://github.com/vercel/next.js/releases/tag/v12.1.1) it came with experimental support for emotion css, so it is time to switch over.

With this PR we're also doing away with any dependencies on [babel-plugin-macros](https://github.com/kentcdodds/babel-plugin-macros) which [isn't really maintained anymore](https://github.com/kentcdodds/babel-plugin-macros/issues/177). 
This means we can't use [@lingui/macro](https://lingui.js.org/ref/macro.html) anymore. This is a bit of a downside, but we found upgrading not too difficult and we migrated the complete codebase in a few hours. And we have to migrate at some point anyways.

We are however very excited about the results of the SWC compiler!

Dev timings:

`next dev` + `babel`: 71,4s seconds
`next dev` + `swc`: 22,7 seconds

On a warm start (after next dev has run once) it is even faster: 6,1 seconds

Build times:

`next build` + `babel`: 8m 12s
`next build` + `swc`: 4m 57s

Modern preset:

We're using a little trick to also force next.js/SWC into a 'modern' mode, this means we don't compile down to ES5, but down to a much more recent version:

| Route     | Old size | New size | Diff    |
| --------- | -------- | -------- | ------- |
| /         | 272kB    | 259.0kB  | -13kB🔥 |
| /\_app    | 196kB    | 190.0kB  | -6kB🔥  |
| /[...url] | 290kB    | 276.0kB  | -14kB🔥 |


The above change isn't officially supported by next.js, but it seems to be working absolutely fine for us. You can apply the [patch](https://github.com/graphcommerce-org/graphcommerce/blob/50188e378b4c77561ebc600958ea11cd114fa61a/patches/next%2B12.1.5.patch) to your own project with the excellent [patch-package](https://www.npmjs.com/package/patch-package)